@@ -5,11 +5,9 @@ buildscript {
         jcenter()
         google()
         maven { url 'https://maven.google.com' }
-        maven { url 'https://maven.fabric.io/public' }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.1'
-        classpath 'io.fabric.tools:gradle:1.+'
+        classpath 'com.android.tools.build:gradle:4.2.1'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -31,7 +29,7 @@ ext {
     // Sdk and tools
     minSdkVersion = 21
     targetSdkVersion = 26
-    compileSdkVersion = 26
+    compileSdkVersion = 28
     buildToolsVersion = '27.0.3'
 
     // App dependencies
